fix(core): moving functions around to eliminate unneeded dependencies
Part of ENG-310
