+++ packages/core/src/command/consolidated/consolidated-create.ts
@@ -1,10 +1,11 @@
-import { Bundle, BundleEntry, DocumentReference } from "@medplum/fhirtypes";
+import { Bundle, BundleEntry } from "@medplum/fhirtypes";
 import { errorToString } from "@metriport/shared";
 import { parseFhirBundle } from "@metriport/shared/medical";
 import dayjs from "dayjs";
 import duration from "dayjs/plugin/duration";
 import { generateAiBriefBundleEntry } from "../../domain/ai-brief/generate";
 import { createConsolidatedDataFilePath } from "../../domain/consolidated/filename";
+import { insertSourceDocumentToAllDocRefMeta } from "../../domain/consolidated/utils";
 import { createFolderName } from "../../domain/filename";
 import { Patient } from "../../domain/patient";
 import { S3Utils, executeWithRetriesS3 } from "../../external/aws/s3";
@@ -12,7 +13,6 @@ import { dangerouslyDeduplicate } from "../../external/fhir/consolidated/dedupli
 import { getDocuments as getDocumentReferences } from "../../external/fhir/document/get-documents";
 import { toFHIR as patientToFhir } from "../../external/fhir/patient/conversion";
 import { buildBundle, buildBundleEntry } from "../../external/fhir/shared/bundle";
-import { getMetriportContent } from "../../external/fhir/shared/extensions/metriport";
 import { capture, executeAsynchronously, out } from "../../util";
 import { Config } from "../../util/config";
 import { controlDuration } from "../../util/race-control";
@@ -207,27 +207,6 @@ export function merge(inputBundle: Bundle) {
       }
       destination.total = destination.entry.length;
       return destination;
-    },
-  };
-}
-
-function insertSourceDocumentToAllDocRefMeta(docRefs: DocumentReference[]): DocumentReference[] {
-  return docRefs.map(insertSourceDocumentToDocRefMeta);
-}
-
-/**
- * Replaces the gibberish string in the `meta.source` generated by the FHIR Server with the title of the source document.
- */
-export function insertSourceDocumentToDocRefMeta<T extends DocumentReference>(docRef: T): T {
-  const metriportContent = getMetriportContent(docRef);
-  const sourceFile = metriportContent?.attachment?.title;
-  if (!sourceFile) return docRef;
-
-  return {
-    ...docRef,
-    meta: {
-      ...docRef.meta,
-      source: sourceFile,
     },
   };
 }
