+++ packages/core/src/domain/consolidated/utils.ts
@@ -1,0 +1,25 @@
+import { DocumentReference } from "@medplum/fhirtypes";
+import { getMetriportContent } from "../../external/fhir/shared/extensions/metriport";
+
+export function insertSourceDocumentToAllDocRefMeta(
+  docRefs: DocumentReference[]
+): DocumentReference[] {
+  return docRefs.map(insertSourceDocumentToDocRefMeta);
+}
+
+/**
+ * Replaces the gibberish string in the `meta.source` generated by the FHIR Server with the title of the source document.
+ */
+export function insertSourceDocumentToDocRefMeta<T extends DocumentReference>(docRef: T): T {
+  const metriportContent = getMetriportContent(docRef);
+  const sourceFile = metriportContent?.attachment?.title;
+  if (!sourceFile) return docRef;
+
+  return {
+    ...docRef,
+    meta: {
+      ...docRef.meta,
+      source: sourceFile,
+    },
+  };
+}
